아파트 상세 정보 조회 로직 수정중
