+++ apps/jibsayo/src/app/api/shared/utils/logger.ts
@@ -1,12 +1,12 @@
-const info = (message: string, data: Record<string, unknown>) => {
+const info = (message: string, data?: Record<string, unknown>) => {
   console.log(`ℹ️ [INFO] ${message}`, data);
 };
 
-const warn = (message: string, data: Record<string, unknown>) => {
+const warn = (message: string, data?: Record<string, unknown>) => {
   console.warn(`⚠️ [WARN] ${message}`, data);
 };
 
-const error = (message: string, data: Record<string, unknown>) => {
+const error = (message: string, data?: Record<string, unknown>) => {
   console.error(`❌ [ERROR] ${message}`, data);
 };
